Add files via upload
Netflix Clone Landing Page

This project is a clone of the Netflix landing page. It showcases my web development skills and attention to detail. The goal was to recreate the look and functionality of the original Netflix website, ensuring it was responsive and visually appealing. This clone demonstrates my proficiency in HTML, CSS, and JavaScript, as well as my ability to follow modern web design trends.